Fix 'jump' command not working
A bug was present where after the Jump function changes
the current line number, if the new line was the last one
in the script, current line was being reset back to the pre-jump one,
cancelling the jump command. "jump if ..." commands were also affected
due to using the Jump function. They were also fixed.